Rename tableheaders.txt to index.tsx and change imports
